requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
langchain-chroma
langchain-huggingface
faiss-cpu
openai
pillow
pymupdf
pypdf
sentence-transformers
chromadb
tqdm
python-dotenv